changed by_header_id/by_mailing_list to by_list_id/by_mailing_list, even though this is a little confusing I upgraded my couchdb by accident and so the python bindings seem to overwrite the by_header_id instead of add a new one.
added from_contacts, to_contacts, cc_contacts, and involves_contacts to the Message model with some changes to the grok_address so it would return tuples instead of just the contact.  The tuples mean that I can know what email the message was sent to, even if we found the contact (who could have multiple identities) via lookup.

The new *_contacts are structured to provide just enough information about the person so a message can be displayed or a full lookup done.
i.e.:   $id : { name : $name, email : $email }

Added some structure to the index.xhtml page so that it moves more toward the look that we want.

Inserted a $("#conversations).empty() function in the onclick handler of the mailing list so there is some instant feedback after clicking.  That might make more sense in an onmouseup() handler instead.